<Append 170106> RUN BROKEN: Compile and Link TaxkBox project. SIGSEG in runtime c'tors. See comment in #3 for more thoughts.
Implemented eIR_USE_FUNCINFO_WORKAROUND [#3] Why is main.cpp still needing references for FuncInfo?
