+++ src/libs/core/base/FuncInfo.cpp
@@ -1,5 +1,7 @@
 #include "FuncInfo.h"
 
+#ifdef eIR_USE_FUNCINFO_WORKAROUND
+
 DEFINE_DATAPROPS(FuncInfo, FuncInfoData)
 void FuncInfo::ctor(void) {;} // for DataProperty default c'tor
 void FuncInfo::dtor(void) {;} // for DataProperty d'tor
@@ -20,3 +22,5 @@ void FuncInfo::parseFuncInfo(void)
 {
     ; // for now
 }
+
+#endif // eIR_USE_FUNCINFO_WORKAROUND
